% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterByOverlap.R
\name{clusterByOverlap}
\alias{clusterByOverlap}
\title{Cluster Overlapping Categories}
\usage{
clusterByOverlap(OG, similarityCut = 0.5, clusteringMethod = "hierarchical")
}
\arguments{
\item{OG}{matrix of overlap}

\item{similarityCut}{minimum similarity required for clustering}

\item{clusteringMethod}{a string, either "hierarchical" or "spectral" to indicate clustering method.}
}
\value{
A list containing
     cluster: cluster assignments,
     plot: ordered plot to visualize clustering,
     plotOrder: ordering for the visualization plot in case you want to try other clusterings,
     overlapMatrix: the original matrix
     diagnosticPlot: A diagnostic plot for the cluster/no cluster decision
}
\description{
First, decides which categories are similar enough to be clustered.  The usual similarity structure is that a few categories are almost identical and then the rest are mainly unique.  Just cluster nearly identical ones
}
\examples{
data("membershipList")

OG <- findCategoryOverlap(membershipList = membershipList)

clusterByOverlap(OG, similarityCut = 0.5, clusteringMethod = "hierarchical")

}
