% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentResult-data.R
\docType{data}
\name{enrichmentResult}
\alias{enrichmentResult}
\title{Proteomic Functional Enrichment Result}
\format{
data.frame
}
\usage{
data(enrichmentResult)
}
\description{
Bacterial proteomic data in response to an antibiotic were submitted to the
\href{https://string-db.org/}{String Database} functional enrichment tool.
This file contains the network neighborhood enrichment data.
}
\examples{
data(enrichmentResult)

}
\references{
Carruthers et al. (2020) Pathogens 18;9(9):766
(\href{https://pubmed.ncbi.nlm.nih.gov/32962036/}{PubMed})
}
\keyword{datasets}
