% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCategoryOverlap.R
\name{findCategoryOverlap}
\alias{findCategoryOverlap}
\title{Category Overlap}
\usage{
findCategoryOverlap(membershipList)
}
\arguments{
\item{membershipList}{Named list where each element is a category and contains a vector of protein identifiers}
}
\value{
Matrix of overlaps, each between 0 - 1.  Matrix names come from make.names(names(membershipList))
}
\description{
Creates an overlap matrix from a list of category members
Overlap is calculated as overlap/(size of the smaller cluster)
}
\examples{
data("membershipList")

findCategoryOverlap(membershipList = membershipList)

}
