% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membershipList-data.R
\docType{data}
\name{membershipList}
\alias{membershipList}
\title{Enriched Category Membership}
\format{
List
}
\usage{
data(membershipList)
}
\description{
Bacterial proteomic data in response to an antibiotic were submitted to the
\href{https://string-db.org/}{String Database} functional enrichment tool.
Enriched network neighborhood output were converted to a list:  each element
contains a vector of gene names found in that category.
}
\examples{
data(membershipList)

}
\references{
Carruthers et al. (2020) Pathogens 18;9(9):766
(\href{https://pubmed.ncbi.nlm.nih.gov/32962036/}{PubMed})
}
\keyword{datasets}
