% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontologyFunctions.R
\name{clusterByOverlap}
\alias{clusterByOverlap}
\title{Cluster Overlapping Categories}
\usage{
clusterByOverlap(OG, similarityCut = 0.5, clusteringMethod = "hierarchical")
}
\arguments{
\item{OG}{matrix of overlap}

\item{similarityCUt}{minimum similarity required for clustering}

\item{method}{a string, either "hierarchical" or "spectral" to indicate clustering method.}
}
\value{
A list containing
     cluster: cluster assignments,
     plot: ordered plot to visualize clustering,
     plotOrder: ordering for the visualization plot in case you want to try other clusterings,
     overlapMatrix: the original matrix
     diagnosticPlot: A diagnostic plot for the cluster/no cluster decision
}
\description{
First, decides which categories are similar enough to be clustered.  The usual similarity structure is that a few categories are almost identical and then the rest are mainly unique.  Just cluster nearly identical ones
}
\examples{
randMembershipList <- list()
set.seed(10)
for(i in 1:27){
  proteinCount <- round(runif(min = 10, max = 150, n = 1))
  randMembershipList[[i]] <- paste0("prot", round(runif(n=proteinCount, min=1, max = 100)))
}
randMembershipList[[28]] <- c(randMembershipList[[27]], randMembershipList[[26]])
randMembershipList[[29]] <- c(randMembershipList[[28]], randMembershipList[[25]])
randMembershipList[[30]] <- c(randMembershipList[[25]], randMembershipList[[27]])
names(randMembershipList) <- paste0("category", 1:30)

OG <- findCategoryOverlap(membershipList = randMembershipList)

clusterByOverlap(OG)

}
