% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontologyFunctions.R
\name{findCategoryOverlap}
\alias{findCategoryOverlap}
\title{Category Overlap}
\usage{
findCategoryOverlap(membershipList)
}
\arguments{
\item{membershipList}{Named list where each element is a category and contains a vector of protein identifiers}
}
\value{
Matrix of overlaps, each between 0 - 1.  Matrix names come from make.names(names(membershipList))
}
\description{
Creates an overlap matrix from a list of category members
Overlap is calculated as overlap/(size of the smaller cluster)
}
\examples{
randMembershipList <- list()
set.seed(10)
for(i in 1:27){
  proteinCount <- round(runif(min = 10, max = 150, n = 1))
  randMembershipList[[i]] <- paste0("prot", round(runif(n=proteinCount, min=1, max = 100)))
}
randMembershipList[[28]] <- c(randMembershipList[[27]], randMembershipList[[26]])
randMembershipList[[29]] <- c(randMembershipList[[28]], randMembershipList[[25]])
randMembershipList[[30]] <- c(randMembershipList[[25]], randMembershipList[[27]])
names(randMembershipList) <- paste0("category", 1:30)

findCategoryOverlap(membershipList = randMembershipList)

}
