% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membershipList-data.R
\docType{data}
\name{membershipList}
\alias{membershipList}
\title{Enriched Category Membership}
\format{
List
}
\usage{
data(membershipList)
}
\description{
A typical enrichment result
}
\examples{
data(membershipList)

}
\references{
none
}
\keyword{datasets}
